Adds Http Exceptions

+++ src/users/users.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, HttpException, HttpStatus } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { User } from './user.entity';
@@ -10,7 +10,17 @@ export class UsersService {
 
     constructor(@InjectRepository(User) private userRepository: Repository<User>) {}
 
-    createUser(user: CreateUserDto) {
+    async createUser(user: CreateUserDto) {
+
+        const userFound = await this.userRepository.findOne({
+            where: {
+                username: user.username
+            }
+        })
+
+        if (userFound) {
+            return new HttpException('User already exists', HttpStatus.CONFLICT)
+        }
         const newUser = this.userRepository.create(user);
         return this.userRepository.save(newUser);
     }
@@ -19,20 +29,54 @@ export class UsersService {
         return this.userRepository.find()
     }
 
-    getUser(id: number) {
-        return this.userRepository.findOne({
+    async getUser(id: number) {
+        const userFound = await this.userRepository.findOne({
             where: {
                 id: id
             }
+        });
+
+        if(!userFound){
+            return new HttpException('User not found', HttpStatus.NOT_FOUND)
+        }
+
+        return userFound;
+    }
+
+    async deleteUser(id: number) {
+        /* const userFound = await this.userRepository.findOne({
+            where: {
+                id
+            }
+        });
+    
+        if(!userFound) {
+            return new HttpException('User not found', HttpStatus.NOT_FOUND);
+        }
+
+        return this.userRepository.delete({ id }); */
+        const result = await this.userRepository.delete({ id });
+
+        if ( result.affected === 0 ) {
+            return new HttpException('User not found', HttpStatus.NOT_FOUND);
+        }
+
+        return result
+    }
+
+    async updateUser(id: number, user: UpdateUserDto) {
+        const userFound = await this.userRepository.findOne({
+            where: {
+                id
+            }
         })
-    }
 
-    deleteUser(id: number) {
-        return this.userRepository.delete({ id });
-    }
+        if (!userFound) {
+            return new HttpException('User not found', HttpStatus.NOT_FOUND);
+        }
 
-    updateUser(id: number, user: UpdateUserDto) {
-        return this.userRepository.update({id: id}, user)
+        const updateUser = Object.assign(userFound, user);
+        return this.userRepository.save(updateUser);
     }
